RavenDB-13688 adding test for encrypted indexing tx limit + minor refactoring to extract encryption setup logic
