+++ test/Tests.Infrastructure/RavenTestBase.cs
@@ -9,6 +9,7 @@ using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Security.Cryptography;
 using System.Security.Cryptography.X509Certificates;
+using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using FastTests.Graph;
@@ -716,6 +717,34 @@ namespace FastTests
             DoNotReuseServer(customSettings);
 
             return serverCertPath;
+        }
+        protected string SetupEncryptedDatabase(out X509Certificate2 adminCert, [CallerMemberName] string caller = null)
+        {
+            var serverCertPath = SetupServerAuthentication();
+            var dbName = GetDatabaseName(caller);
+            adminCert = AskServerForClientCertificate(serverCertPath, new Dictionary<string, DatabaseAccess>(), SecurityClearance.ClusterAdmin);
+
+            var buffer = new byte[32];
+            using (var rand = RandomNumberGenerator.Create())
+            {
+                rand.GetBytes(buffer);
+            }
+
+            var base64Key = Convert.ToBase64String(buffer);
+
+            // sometimes when using `dotnet xunit` we get platform not supported from ProtectedData
+            try
+            {
+                ProtectedData.Protect(Encoding.UTF8.GetBytes("Is supported?"), null, DataProtectionScope.CurrentUser);
+            }
+            catch (PlatformNotSupportedException)
+            {
+                // so we fall back to a file
+                Server.ServerStore.Configuration.Security.MasterKeyPath = GetTempFileName();
+            }
+
+            Server.ServerStore.PutSecretKey(base64Key, dbName, true);
+            return dbName;
         }
 
         public class Options
